refactor(server): extract shared auth and error plugins
