+++ apps/server/package.json
@@ -14,11 +14,12 @@
     "start": "bun run dist/index.mjs"
   },
   "dependencies": {
-    "@anilog/api": "workspace:*",
     "@anilog/auth": "workspace:*",
+    "@anilog/contracts": "workspace:*",
+    "@anilog/domain": "workspace:*",
     "@elysiajs/cors": "^1.3.3",
     "dotenv": "catalog:",
-    "elysia": "^1.4.21"
+    "elysia": "^1.4.26"
   },
   "devDependencies": {
     "@anilog/config": "workspace:*",
