+++ apps/server/src/plugins/auth.plugin.ts
@@ -1,0 +1,18 @@
+import { Elysia } from "elysia";
+import { auth } from "@anilog/auth";
+import { unauthorizedError } from "@anilog/domain";
+
+export const authPlugin = (app: Elysia) =>
+  app.derive(async ({ request }) => {
+    const session = await auth.api.getSession({ headers: request.headers });
+    const userId = session?.user?.id;
+
+    if (!userId) {
+      throw unauthorizedError("User not authenticated");
+    }
+
+    return {
+      session,
+      userId,
+    };
+  });
