+++ apps/server/src/routes/admin.route.ts
@@ -1,6 +1,5 @@
 import { Elysia, t } from "elysia";
-import { AdminService, UserService, forbiddenError, unauthorizedError } from "@anilog/api";
-import { auth } from "@anilog/auth";
+import { AdminService } from "@anilog/domain";
 import {
   AdminUsersQuery,
   AdminStatsDto,
@@ -8,27 +7,14 @@ import {
   HeroCurationDto,
   SetUserAdminStatusBody,
   SetUserAdminStatusDto,
+  HeroCurationParams,
   UserParams,
   UpdateHeroCurationBody,
-} from "@anilog/api";
-
-const adminMiddleware = (app: Elysia) =>
-  app.derive(async ({ request }) => {
-    const session = await auth.api.getSession({ headers: request.headers });
-    if (!session?.user?.id) {
-      throw unauthorizedError("User not authenticated");
-    }
-
-    const { isAdmin } = await UserService.getAdminStatus(session.user.id);
-    if (!isAdmin) {
-      throw forbiddenError("Forbidden");
-    }
-
-    return { adminUserId: session.user.id };
-  });
+} from "@anilog/contracts";
+import { adminPlugin } from "../plugins/admin.plugin";
 
 export const adminRoutes = new Elysia({ prefix: "/admin" })
-  .use(adminMiddleware)
+  .use(adminPlugin)
   .get(
     "/stats",
     async () => {
@@ -74,7 +60,7 @@ export const adminRoutes = new Elysia({ prefix: "/admin" })
       return AdminService.updateHeroCuration(params.id, body);
     },
     {
-      params: t.Object({ id: t.Numeric() }),
+      params: HeroCurationParams,
       body: UpdateHeroCurationBody,
       response: HeroCurationDto,
     },
