+++ apps/web/package.json
@@ -19,7 +19,7 @@
     "better-auth": "catalog:",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "elysia": "^1.4.21",
+    "elysia": "^1.4.26",
     "framer-motion": "^12.34.0",
     "lucide-react": "^0.564.0",
     "next": "^16.0.10",
